backend/requirements.txt
alembic==1.15.2
celery[redis]==5.3.*
fastapi==0.115.12
pydantic==2.11.3
pydantic_settings==2.8.1
slowapi==0.1.9
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
structlog==25.2.0
uvicorn==0.34.1
email-validator==2.2.0
asyncpg==0.30.0
itsdangerous==2.2.0
psycopg2-binary==2.9.9
openai==1.84.0
boto3==1.37.22
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
pytest-mock==3.14.0
redis==5.0.*
syllables==1.0.7
black==25.1.0
isort==5.13.2
mypy==1.16.1
ruff==0.12.2
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
# Type stubs for mypy
types-python-jose==3.5.0.20250531
types-passlib==1.7.7.20250602
greenlet==3.2.3
